pyproject.toml
# pyproject.toml

[build-system]
requires      = ["hatchling",]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
exclude = ["docs*", "test*", "__pycache__", ".vscode", "datasets_to_merge",
"merged_dataset*", "runs", 
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[project]
name = "yolov8_dataset_merger"
version = "1.0.0"
description = "Merge Datasets of different formats without conflicts"
readme = "README.md"
authors = [{ name = "Frederic Pepin", email = "fredericpepin3@gmail.com" }]
license = { text = "MIT License" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
keywords = ["formatting-datasets", "datasets", "yolov8"]
dependencies = [
'alabaster==0.7.16',
'Babel==2.15.0',
'build==1.2.2',
'certifi==2024.6.2',
'charset-normalizer==3.3.2',
'colorama==0.4.6',
'contourpy==1.2.1',
'cycler==0.12.1',
'defusedxml==0.7.1',
'docutils==0.20.1',
'filelock==3.15.4',
'fonttools==4.53.0',
'fsspec==2024.6.0',
'idna==3.7',
'imagesize==1.4.1',
'intel-openmp==2021.4.0',
'Jinja2==3.1.4',
'kiwisolver==1.4.5',
'MarkupSafe==2.1.5',
'matplotlib==3.9.0',
'mkl==2021.4.0',
'mpmath==1.3.0',
'networkx==3.3',
'numpy==1.26.4',
'opencv-python==4.10.0.84',
'opencv-python-headless==4.10.0.84',
'packaging==24.1',
'pandas==2.2.2',
'pillow==10.3.0',
'pockets==0.9.1',
'psutil==6.0.0',
'py-cpuinfo==9.0.0',
'pycocotools==2.0.8',
'Pygments==2.18.0',
'pyparsing==3.1.2',
'pyproject_hooks==1.2.0',
'python-dateutil==2.9.0.post0',
'pytz==2024.1',
'PyYAML==6.0.1',
'requests==2.32.3',
'scipy==1.14.0',
'seaborn==0.13.2',
'setuptools==75.1.0',
'six==1.16.0',
'snowballstemmer==2.2.0',
'Sphinx==7.3.7',
'sphinx-rtd-theme==2.0.0',
'sphinxcontrib-applehelp==1.0.8',
'sphinxcontrib-devhelp==1.0.6',
'sphinxcontrib-htmlhelp==2.0.5',
'sphinxcontrib-jquery==4.1',
'sphinxcontrib-jsmath==1.0.1',
'sphinxcontrib-napoleon==0.7',
'sphinxcontrib-qthelp==1.0.7',
'sphinxcontrib-serializinghtml==1.1.10',
'supervision==0.21.0',
'sympy==1.12.1',
'tbb==2021.13.0',
'tia-openness-api-client==0.1.0',
'torch==2.3.1',
'torchvision==0.18.1',
'tqdm==4.66.4',
'typing_extensions==4.12.2',
'tzdata==2024.1',
'ultralytics==8.2.42',
'ultralytics-thop==2.0.0',
'urllib3==2.2.2',
]
requires-python = ">=3.12"

[project.urls]
Documentation = "https://readthedocs.org"
Repository = "https://github.com/fredericpepin/yolov8-dataset-merger"
